+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-bludgeon: Added `-T excessive-script-nodes` which detects and removes
   script node tables that exceed 19001 (tags that do this potentially crash
   stock Halo PC)
+- invader-index: Added feedback for if the command was successful
+- invader-resource: Added feedback for if the command was successful
 
 ### Changed
 - invader: Fixed segfault when querying dependencies for various tools